move valueOrThrow to general utils
