+++ src/util.ts
@@ -1,7 +1,8 @@
 import * as debug from 'debug';
 import * as minimist from 'minimist';
+import * as t from 'io-ts';
+import { ThrowReporter } from 'io-ts/lib/ThrowReporter';
 import { Args } from './model';
-import { valueOrThrow } from './scripts/webpack/util';
 
 debug.enable('scriptoni:*');
 
@@ -9,7 +10,6 @@ export const logger = {
   bin: debug('scriptoni:bin'),
   metarpheus: debug('scriptoni:metarpheus'),
   metarpheusDiff: debug('scriptoni:metarpheus-diff'),
-  lint: debug('scriptoni:lint'),
   lintStyle: debug('scriptoni:lint-style'),
   webpack: debug('scriptoni:webpack'),
   prettier: debug('scriptoni:prettier')
@@ -20,6 +20,17 @@ const defaultArgs: Partial<Args> = {
   metarpheusConfig: 'metarpheus-ts-config.js',
   c: './config'
 };
+
+export function valueOrThrow<O, I>(iotsType: t.Type<O, I>, value: I): O {
+  const validatedValue = iotsType.decode(value);
+
+  if (validatedValue.isLeft()) {
+    ThrowReporter.report(validatedValue);
+    return undefined as never;
+  } else {
+    return validatedValue.value;
+  }
+}
 
 export const getArgs = (): Args => {
   const args = {
